requirements.txt
altgraph==0.17.4
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
importlib-metadata==4.13.0
Jinja2==3.1.4
lxml==5.2.1
MarkupSafe==2.1.5
nidaqmx==0.8.0
numpy>=1.22,<1.26
opcua==0.98.13
opencv-python==4.10.0.84
packaging==24.0
pandas==1.3.5
pefile==2024.8.26
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2024.7
pypylon==3.0.1
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
pyqtgraph==0.12.4
pyserial==3.5
PySide2==5.15.2.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.1
pywin32-ctypes==0.2.3
qt-material==2.14
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
shiboken2==5.15.2.1
six==1.16.0
typing_extensions==4.7.1
zipp==3.15.0
